Build subset without COD present in FFS

+++ cm_transport.h
@@ -58,12 +58,12 @@ typedef CMTaskHandle (*CMTransport_add_period_task)(CManager cm,
 typedef void (*CMTransport_remove_periodic)(CMTaskHandle cmt);
 typedef void (*CMTransport_add_poll)(CManager cm, CMPollFunc func,
 				     void *client_data);
-typedef CMbuffer (*CMTransport_get_data_buffer)(CManager cm, int length);
+typedef CMbuffer (*CMTransport_get_data_buffer)(CManager cm, size_t length);
 typedef void (*CMTransport_return_data_buffer)(CManager cm, CMbuffer cmb);
 typedef void (*CMTransport_connection_close)(CMConnection conn);
 typedef void *(*CMTransport_get_transport_data)(CMConnection conn);
 typedef void (*CMTransport_action_pending_write)(CMConnection conn);
-typedef CMbuffer (*CMTransport_create_data_buffer)(CManager cm, void *buffer, int length);
+typedef CMbuffer (*CMTransport_create_data_buffer)(CManager cm, void *buffer, size_t length);
 typedef int (*CMTransport_modify_global_lock)(CManager cm, const char *file, int line);
 typedef void (*CMTransport_add_buffer_to_pending_queue)(CManager cm, CMConnection conn, CMbuffer buf, long length);
 typedef void (*CMTransport_cond_wait_CM_lock)(CManager cm, void *cond, char *file, int line);
@@ -123,20 +123,20 @@ typedef attr_list (*CMTransport_listen_func)(CManager cm,
 					     attr_list listen_info);
 typedef void *(*CMTransport_read_block_func)(CMtrans_services svc,
 					     void *conn_data,
-					     int *actual, int *offset);
+					     size_t *actual, size_t *offset);
 typedef int (*CMTransport_read_to_buffer_func)(CMtrans_services svc,
 					       void *conn_data,
 					       void *buffer,
-					       int len, int block_flag);
+					       size_t len, int block_flag);
 typedef int (*CMTransport_writev_func)(CMtrans_services svc,
 				       void *transport_data,
-				       void *buffer, int len,
+				       void *buffer, size_t len,
 				       attr_list attrs);
 
 typedef void (*CMcompletion_notify_func)(void *client_data);
 typedef int (*CMTransport_writev_complete_notify_func)(CMtrans_services svc,
 						       void *transport_data,
-						       void *buffer, int len,
+						       void *buffer, size_t len,
 						       attr_list attrs, CMcompletion_notify_func func,
 						       void *client_data);
 typedef void (*CMTransport_shutdown_conn_func)(CMtrans_services svc,
